test: verify eligibility flag is set to true for add participant
